refactor: created GlobalAppDataConfig file
removed enum type and interfaces from AppConstant.ts and created GlobalAppDataConfig for enum.
