+++ src/utility/AppInterfaces.ts
@@ -45,3 +45,41 @@ export interface BioDeviceConfig {
   code: string;
   name: string;
 }
+
+export interface ImgeSizeConfig {
+  size: number;
+}
+
+export interface AccessTokenRequest {
+  grant_type: string;
+  client_id: string;
+  client_secret: string;
+  username?: string;
+  password?: string;
+  refresh_token?: string;
+}
+
+export interface ORPAuthRequestUsed {
+  transaction_type: string;
+  customer_type: string;
+  make: string;
+  model: string;
+  year: string;
+  trim: string;
+  kms_driven: string;
+  city: string;
+  noOfOwners: string;
+}
+
+export interface RCNumberRequest {
+  rc_number: string;
+}
+
+export interface ORPAuthRequestNew {
+  category: string;
+  make?: string;
+  model?: string;
+  year?: string;
+  trim?: string;
+  city?: string;
+}
